kernel: Fix 'mult' overflow if using 10 Hz
Change 'mult' to 64-bit (u64) to fix overflow warning if using 10 Hz jiffies.

./include/vdso/jiffies.h:9:19: warning: unsigned conversion from 'long int' to 'unsigned int' changes value from '6400000000' to '2105032704' [-Woverflow]
    9 | #define TICK_NSEC ((NSEC_PER_SEC+HZ/2)/HZ)
      |                   ^
kernel/time/jiffies.c:38:35: note: in expansion of macro 'TICK_NSEC'
   38 |         .mult                   = TICK_NSEC << JIFFIES_SHIFT, /* details above */
      |                                   ^~~~~~~~~

+++ include/linux/clockchips.h
@@ -104,7 +104,7 @@ struct clock_event_device {
 	ktime_t			next_event;
 	u64			max_delta_ns;
 	u64			min_delta_ns;
-	u32			mult;
+	u64			mult;
 	u32			shift;
 	enum clock_event_state	state_use_accessors;
 	unsigned int		features;
